Use an internal route
Under load test conditions, traffic to a public routes is sent via the PaaS gorouter. This causes traffic to be amplified within the platform which does not accurately reflect production conditions.

We can use an internal route which routes traffic directly the the app via internal hostname. Ideally this app should be hosted outside of the isolation segment or PaaS to simulate real-world remote network conditions - this can be addressed at a later point.