Use internal paas networking for staging
When running load tests recently, I noticed an unusual distribution of
requests from here to the API: 90% of the requests were being sent to
only a few of the API instances, causing uneven load that led to the
load test results being somewhat inconsistent/unreliable.

We aren't exactly sure why this has been happening, but switching over
to using an internal route for the API has alleviated the problem and
left us with a clean distribution of requests across all API instances.