Update clients to optionally pass through the ECS header
We want both SMS clients to be able to pass through the ECS header
("x-notify-ecs-origin") so that they can be used to stub requests from
the ECS apps.

This change passes the header through if the `USE_ECS_APPS` environment
variable has a value of `"true"`. If the variable is not set or has any
other value the header will not be passed.

Passing the header means that we need to change the client methods that
are used to allow this.